feat: add main routing and not found page
