+++ class-components/src/components/pages/notFoundPage/NotFoundPage.tsx
@@ -1,0 +1,8 @@
+const NotFoundPage = () => {
+  return (
+    <div>
+      <img style={{ display: 'block', width: '100%' }} src="/assets/404.webp" />
+    </div>
+  );
+};
+export default NotFoundPage;
