+++ class-components/src/components/pagination/pagination.css
@@ -18,10 +18,6 @@
   background-color: rgb(128,128,128,0.1);
 }
 
-.pagination__item:active {
-  background-color: rgb(128,128,128,0.1);
-}
-
 .pagination__link {
   text-decoration: none;
 }
